servo: Merge #600 - Rewrite reftest harness and add basic example reftest (from metajack:new-reftest); r=pcwalton
This does not port the existing src/test/html/ref tests to the new framework, as it appears to me that they aren't really reftests in the sense of Gecko's reftest. This new driver uses the Gecko methodology.

Currently this will pop a window for each test due to not having a headless driver yet, and #570 means that servo segfaults when it shuts down so we can't check the exit status.

There's plenty to improve in the future, but this should get us started.

Source-Repo: https://github.com/servo/servo
Source-Revision: 41f7109c63c5b4558ea9ab1fc0c5c9f4a8294357