servo: Merge #968 - Removing unnecessary constant window title update (from luisbg:master); r=larsbergstrom
update_window_title() is called constantly in the compositor loop. This function
always changes the title to "Servo" when in idle state so it is spamming the X
server with constant changes. But this isn't necessary because updating the
title is taken care of when the Rendering or Ready state are changed in
set_ready_state() and set_render_state().

Fixes #830

Happy Software Freedom Day

Source-Repo: https://github.com/servo/servo
Source-Revision: 0096d119f084ea65ee69cbd5beda994857601eec