servo: Merge #1413 - Adjust PNG encoder color type to match what's read from gl2 (from dhedlund:png_color_type); r=pcwalton
This should fix images produced by the -o flag.

The gl2::read_pixels call was asking for pixel data in RGB8 but was then telling the PNG encoder that the data was RGBA8.  The value was originally set correctly, but was changed when support was added for drawing transparent PNGs.

We could instead adjust the gl2::read_pixels call to return RGBA8 so we can write transparent screenshots, but I'm not sure if there's a use case for that.

Source-Repo: https://github.com/servo/servo
Source-Revision: 44404766da6a551d5f24ebffa290ae9a6f5e14a5