servo: Merge #5245 - Make IOCompositor only respond to the first Quit event (from nnethercote:only-one-quit-event); r=glennw
This fixes #5234, in that the huge memory spike disappears. It still takes ~15 seconds for the window to actually disappear after that first `Quit` event is received by the IOCompositor. Maybe that's a pre-existing problem.

There may be better ways to do this, like handling it on the sending side (i.e. within glutin) instead of the receiving side. I just did it this way because it seemed like the easiest thing.

Source-Repo: https://github.com/servo/servo
Source-Revision: 2281bca89252d8e37771ce2446fb5871a69903d0